[CABMAN] Move CCFDATAStorage class to a separate file
